fix : minmumPrice int 타입으로 교체, StoreServiceTest 수정 StoreServiceTest의 Store를 영업 준비한다를 고정된 시간으로 테스트하게 코드가 수정됐습니다.

+++ src/main/java/com/musa3team/devout/domain/store/dto/StoreRequestDto.java
@@ -20,6 +20,6 @@ public class StoreRequestDto {
     private LocalTime open_time;
     @ValidTenMinuteInterval
     private LocalTime close_time;
-    private Long minimum_price;
+    private int minimum_price;
     private StoreCategory category;
 }
